+++ components/src/main/java/org/patternfly/component/BaseComponent.java
@@ -54,7 +54,12 @@ public abstract class BaseComponent<E extends HTMLElement, B extends TypedBuilde
     protected BaseComponent(ComponentType componentType, E element) {
         this.componentType = requireNonNull(componentType, "component type required");
         this.element = requireNonNull(element, "element required");
-        ouia(element, componentType.id, componentType.componentName);
+        ouia(element, componentType.componentName);
+    }
+
+    public B ouiaId(String id) {
+        ouia(element(), id, componentType().componentName);
+        return that();
     }
 
     @Override